requirements.txt
numpy
torch==1.8.1
torchvision
pillow
opencv-python